+++ app/index.php
@@ -1,0 +1,6 @@
+<?php
+/**
+ * Silence is golden.
+ *
+ * @package nested-ordered-lists-for-block-editor
+ */