+++ uninstall.php
@@ -10,10 +10,19 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
+use nestedOrderedLists\Transient;
+use nestedOrderedLists\Transients;
+
 // if uninstall.php is not called by WordPress, die.
 if ( ! defined( 'WP_UNINSTALL_PLUGIN' ) ) {
 	die;
 }
+
+// set name for transient list.
+const NOLG_TRANSIENTS_LIST = 'nolg_transients';
+
+// embed necessary files.
+require_once __DIR__ . '/vendor/autoload.php';
 
 // delete the content of all taxonomies.
 global $wpdb;
@@ -45,3 +54,17 @@ $wpdb->delete( $wpdb->term_taxonomy, array( 'taxonomy' => 'nolg_icon_set' ), arr
 
 // cleanup options.
 delete_option( 'nolg_icon_set_children' );
+
+// delete transients.
+foreach ( Transients::get_instance()->get_transients() as $transient ) {
+	// bail if object is not a transient object.
+	if ( ! $transient instanceof Transient ) {
+		continue;
+	}
+
+	// delete the dismiss setting.
+	$transient->delete_dismiss();
+
+	// delete the transient.
+	$transient->delete();
+}